feat: support swipe interaction.
