+++ src/interaction/index.js
@@ -2,6 +2,7 @@ module.exports = {
   Interaction: require('./base'),
   PieSelect: require('./pie-select'),
   IntervalSelect: require('./interval-select'),
+  Swipe: require('./swipe'),
   Pan: require('./pan'),
   Pinch: require('./pinch')
 };
