+++ src/interaction/pan.js
@@ -1,18 +1,8 @@
 const Util = require('../util/common');
-const Helper = require('./helper');
 const Interaction = require('./base');
 const Chart = require('../chart/chart');
 const FilterPlugin = require('../plugin/filter');
-const DAY_TIMESTAMPS = 86400000;
-
-const TOUCH_EVENTS = [
-  'touchstart',
-  'touchmove',
-  'touchend',
-  'touchStart',
-  'touchMove',
-  'touchEnd'
-];
+const MoveMixin = require('./mixin/move');
 
 class Pan extends Interaction {
   getDefaultCfg() {
@@ -78,6 +68,8 @@ class Pan extends Interaction {
         self.limitRange = {};
       }
     }]);
+
+    Util.mix(this, MoveMixin);
   }
 
   start(e) {
@@ -87,12 +79,12 @@ class Pan extends Interaction {
     if (e.type === 'touchstart' || e.type === 'touchStart') {
       this.lastPoint = e.touches[0];
     }
-    this._handlePan(e);
+    this._handleMove(e);
   }
 
   process(e) {
     if (this.pressed) return;
-    this._handlePan(e);
+    this._handleMove(e);
   }
 
   end() {
@@ -117,156 +109,6 @@ class Pan extends Interaction {
     const center = e.center || e.touches[0];
     this.chart.tooltip(true);
     this.chart.showTooltip(center);
-  }
-
-  _handlePan(e) {
-    const { currentDeltaX, currentDeltaY, lastPoint } = this;
-    let deltaX;
-    let deltaY;
-    if (TOUCH_EVENTS.indexOf(e.type) !== -1) {
-      const currentPoint = e.touches[0];
-      deltaX = currentPoint.x - lastPoint.x;
-      deltaY = currentPoint.y - lastPoint.y;
-      this.lastPoint = currentPoint;
-    } else if (currentDeltaX !== null && currentDeltaY !== null) {
-      this.panning = true;
-      deltaX = e.deltaX - currentDeltaX;
-      deltaY = e.deltaY - currentDeltaY;
-      this.currentDeltaX = e.deltaX;
-      this.currentDeltaY = e.deltaY;
-    }
-
-    if (!Util.isNil(deltaX) || !Util.isNil(deltaY)) {
-      const lastTimestamp = this._timestamp;
-      const now = +new Date();
-      if ((now - lastTimestamp) > 16) {
-        this._doPan(deltaX, deltaY);
-        this._timestamp = now;
-      }
-    }
-  }
-
-  _doPan(deltaX, deltaY) {
-    const self = this;
-    const { mode, chart, limitRange } = self;
-    const coord = chart.get('coord');
-    const { start, end } = coord;
-    const data = chart.get('data');
-    if (Helper.directionEnabled(mode, 'x') && deltaX !== 0) {
-      const xScale = chart.getXScale();
-      const xField = xScale.field;
-      if (!limitRange[xField]) {
-        limitRange[xField] = Helper._getLimitRange(data, xScale);
-      }
-
-      const coordWidth = end.x - start.x;
-
-      if (xScale.isCategory) {
-        self._panCatScale(xScale, deltaX, coordWidth);
-      } else if (xScale.isLinear) {
-        self._panLinearScale(xScale, deltaX, coordWidth, 'x');
-      }
-      const xDef = Helper.getColDef(chart, xField);
-      this.xRange = Helper._getFieldRange(xDef, limitRange[xField], xScale.type);
-    }
-
-    if (Helper.directionEnabled(mode, 'y') && deltaY !== 0) {
-      const coordHeight = start.y - end.y;
-      const yScales = chart.getYScales();
-      Util.each(yScales, yScale => {
-        const yField = yScale.field;
-        if (!limitRange[yField]) {
-          limitRange[yField] = Helper._getLimitRange(data, yScale);
-        }
-
-        yScale.isLinear && self._panLinearScale(yScale, deltaY, coordHeight, 'y');
-      });
-      const yDef = Helper.getColDef(chart, yScales[0].field);
-      this.yRange = Helper._getFieldRange(yDef, limitRange[yScales[0].field], yScales[0].type);
-    }
-    chart.repaint();
-  }
-
-  _panLinearScale(scale, delta, range, flag) {
-    const { field, min, max } = scale;
-    const limitRange = this.limitRange;
-
-    if (min === limitRange[field].min && max === limitRange[field].max) return;
-
-    const chart = this.chart;
-    const ratio = delta / range;
-    const panValue = ratio * (max - min);
-    let newMax = flag === 'x' ? max - panValue : max + panValue;
-    let newMin = flag === 'x' ? min - panValue : min + panValue;
-    if (limitRange[field] && !Util.isNil(limitRange[field].min) && newMin <= limitRange[field].min) {
-      newMin = limitRange[field].min;
-      newMax = (max - min) + newMin;
-    }
-    if (limitRange[field] && !Util.isNil(limitRange[field].max) && newMax >= limitRange[field].max) {
-      newMax = limitRange[field].max;
-      newMin = newMax - (max - min);
-    }
-    const colDef = Helper.getColDef(chart, field);
-    chart.scale(field, Util.mix({}, colDef, {
-      min: newMin,
-      max: newMax,
-      nice: false
-    }));
-  }
-
-  _panCatScale(scale, delta, range) {
-    const { type, field, values, ticks } = scale;
-    const chart = this.chart;
-    const colDef = Helper.getColDef(chart, field);
-
-    const originValues = this.limitRange[field];
-    const lastValueIndex = originValues.length - 1;
-    const currentLength = values.length;
-    const speed = this.speed || 1;
-    const step = range / (currentLength * speed);
-
-    const firstIndex = originValues.indexOf(values[0]);
-    const lastIndex = originValues.indexOf(values[currentLength - 1]);
-    let minIndex = firstIndex;
-    let maxIndex = lastIndex;
-
-    const ratio = Math.abs(delta / range);
-    const panStep = this.step || Math.max(1, parseInt(ratio * currentLength));
-
-    this._panCumulativeDelta += delta;
-    minIndex = this._panCumulativeDelta > step ?
-      Math.max(0, minIndex - panStep) :
-      (this._panCumulativeDelta < -step ? Math.min(lastValueIndex - currentLength + 1, minIndex + panStep) : minIndex);
-
-    maxIndex = Math.min(lastValueIndex, minIndex + currentLength - 1);
-
-    if (minIndex === firstIndex && maxIndex === lastIndex) {
-      return null;
-    }
-
-    const newValues = originValues.slice(minIndex, maxIndex + 1);
-    let newTicks = null;
-    if (type === 'timeCat') {
-      const tickGap = ticks.length > 2 ? ticks[1] - ticks[0] : DAY_TIMESTAMPS;
-
-      if (this._panCumulativeDelta > step) {
-        for (let i = ticks[0] - tickGap; i >= newValues[0]; i -= tickGap) {
-          ticks.unshift(i);
-        }
-      } else if (this._panCumulativeDelta < -step) {
-        for (let i = ticks[ticks.length - 1] + tickGap; i <= newValues[newValues.length - 1]; i += tickGap) {
-          ticks.push(i);
-        }
-      }
-
-      newTicks = ticks;
-    }
-
-    chart.scale(field, Util.mix({}, colDef, {
-      values: newValues,
-      ticks: newTicks
-    }));
-    this._panCumulativeDelta = minIndex !== firstIndex ? 0 : this._panCumulativeDelta;
   }
 }
