+++ test/unit/interaction/index-spec.js
@@ -3,6 +3,6 @@ const Interaction = require('../../../src/interaction/');
 
 describe('Current Interactions', () => {
   it('Type of Interactions', () => {
-    expect(Interaction).to.have.all.keys('Interaction', 'PieSelect', 'IntervalSelect', 'Pan', 'Pinch');
+    expect(Interaction).to.have.all.keys('Interaction', 'PieSelect', 'IntervalSelect', 'Pan', 'Pinch', 'Swipe');
   });
 });
